requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
python-decouple>=3.8
whitenoise>=6.5.0
gunicorn>=21.2.0
